+++ packages/sources/mobula-state/src/index.ts
@@ -1,13 +1,15 @@
 import { expose, ServerInstance } from '@chainlink/external-adapter-framework'
-import { Adapter } from '@chainlink/external-adapter-framework/adapter'
+import { PriceAdapter } from '@chainlink/external-adapter-framework/adapter'
 import { config } from './config'
+import includes from './config/includes.json'
 import { fundingRate, price } from './endpoint'
 
-export const adapter = new Adapter({
+export const adapter = new PriceAdapter({
   defaultEndpoint: price.name,
   name: 'MOBULA_STATE',
   config,
   endpoints: [price, fundingRate],
+  includes,
 })
 
 export const server = (): Promise<ServerInstance | undefined> => expose(adapter)