+++ packages/composites/bsol-price/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chainlink/bsol-price-adapter
 
+## 2.2.136
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @chainlink/token-allocation-adapter@1.14.4
+
 ## 2.2.135
 
 ### Patch Changes