+++ packages/composites/crypto-volatility-index/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chainlink/crypto-volatility-index-adapter
 
+## 1.4.32
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @chainlink/token-allocation-adapter@1.14.4
+
 ## 1.4.31
 
 ### Patch Changes