+++ packages/composites/defi-pulse/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chainlink/defi-pulse-adapter
 
+## 1.2.137
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @chainlink/token-allocation-adapter@1.14.4
+
 ## 1.2.136
 
 ### Patch Changes