+++ packages/composites/multi-address-list/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @chainlink/multi-address-list-adapter
 
+## 1.1.0
+
+### Minor Changes
+
+- [#4348](https://github.com/smartcontractkit/external-adapters-js/pull/4348) [`8720c2a`](https://github.com/smartcontractkit/external-adapters-js/commit/8720c2a6f5c759e9503223c7f39e11a228761f18) Thanks [@Subarna-Singh](https://github.com/Subarna-Singh)! - Update EA Framework Version and Integration test Snapshot
+
+### Patch Changes
+
+- Updated dependencies [[`d61b0e6`](https://github.com/smartcontractkit/external-adapters-js/commit/d61b0e699224ea2b393d717f2cfb4376f921444c), [`8720c2a`](https://github.com/smartcontractkit/external-adapters-js/commit/8720c2a6f5c759e9503223c7f39e11a228761f18)]:
+  - @chainlink/anchorage-adapter@2.2.0
+
 ## 1.0.24
 
 ### Patch Changes