+++ packages/composites/set-token-index/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chainlink/set-token-index-adapter
 
+## 3.0.45
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @chainlink/token-allocation-adapter@1.14.4
+
 ## 3.0.44
 
 ### Patch Changes